+++ hahomematic/client/__init__.py
@@ -42,7 +42,7 @@ from hahomematic.const import (
     SystemInformation,
     SystemVariableData,
 )
-from hahomematic.decorators import measure_execution_time, service
+from hahomematic.decorators import async_inspector, measure_execution_time
 from hahomematic.exceptions import BaseHomematicException, ClientException, NoConnectionException
 from hahomematic.model.device import Device
 from hahomematic.model.support import convert_value
@@ -298,7 +298,7 @@ class Client(ABC):
     async def fetch_device_details(self) -> None:
         """Fetch names from backend."""
 
-    @service(re_raise=False, no_raise_return=False)
+    @async_inspector(re_raise=False, no_raise_return=False)
     async def is_connected(self) -> bool:
         """
         Perform actions required for connectivity check.
@@ -404,7 +404,7 @@ class Client(ABC):
                     return device
         return None
 
-    @service(re_raise=False)
+    @async_inspector(re_raise=False)
     async def get_device_description(self, device_address: str) -> DeviceDescription | None:
         """Get device descriptions from CCU / Homegear."""
         try:
@@ -417,7 +417,7 @@ class Client(ABC):
             _LOGGER.warning("GET_DEVICE_DESCRIPTIONS failed: %s [%s]", ex.name, reduce_args(args=ex.args))
         return None
 
-    @service()
+    @async_inspector()
     async def add_link(self, sender_address: str, receiver_address: str, name: str, description: str) -> None:
         """Return a list of links."""
         try:
@@ -427,7 +427,7 @@ class Client(ABC):
                 f"ADD_LINK failed with for: {sender_address}/{receiver_address}/{name}/{description}: {reduce_args(args=ex.args)}"
             ) from ex
 
-    @service()
+    @async_inspector()
     async def remove_link(self, sender_address: str, receiver_address: str) -> None:
         """Return a list of links."""
         try:
@@ -437,7 +437,7 @@ class Client(ABC):
                 f"REMOVE_LINK failed with for: {sender_address}/{receiver_address}: {reduce_args(args=ex.args)}"
             ) from ex
 
-    @service()
+    @async_inspector()
     async def get_link_peers(self, address: str) -> tuple[str, ...] | None:
         """Return a list of link pers."""
         try:
@@ -445,7 +445,7 @@ class Client(ABC):
         except BaseHomematicException as ex:
             raise ClientException(f"GET_LINK_PEERS failed with for: {address}: {reduce_args(args=ex.args)}") from ex
 
-    @service()
+    @async_inspector()
     async def get_links(self, address: str, flags: int) -> dict[str, Any]:
         """Return a list of links."""
         try:
@@ -453,7 +453,7 @@ class Client(ABC):
         except BaseHomematicException as ex:
             raise ClientException(f"GET_LINKS failed with for: {address}: {reduce_args(args=ex.args)}") from ex
 
-    @service()
+    @async_inspector()
     async def get_metadata(self, address: str, data_id: str) -> dict[str, Any]:
         """Return the metadata for an object."""
         try:
@@ -463,7 +463,7 @@ class Client(ABC):
                 f"GET_METADATA failed with for: {address}/{data_id}: {reduce_args(args=ex.args)}"
             ) from ex
 
-    @service()
+    @async_inspector()
     async def set_metadata(self, address: str, data_id: str, value: dict[str, Any]) -> dict[str, Any]:
         """Write the metadata for an object."""
         try:
@@ -473,7 +473,7 @@ class Client(ABC):
                 f"SET_METADATA failed with for: {address}/{data_id}/{value}: {reduce_args(args=ex.args)}"
             ) from ex
 
-    @service(log_level=logging.NOTSET)
+    @async_inspector(log_level=logging.NOTSET)
     async def get_value(
         self,
         channel_address: str,
@@ -499,7 +499,7 @@ class Client(ABC):
                 f"GET_VALUE failed with for: {channel_address}/{parameter}/{paramset_key}: {reduce_args(args=ex.args)}"
             ) from ex
 
-    @service(measure_performance=True)
+    @async_inspector(measure_performance=True)
     async def _set_value(
         self,
         channel_address: str,
@@ -592,7 +592,7 @@ class Client(ABC):
             ):
                 data_point.write_temporary_value(value=value)
 
-    @service(re_raise=False, no_raise_return=set())
+    @async_inspector(re_raise=False, no_raise_return=set())
     async def set_value(
         self,
         channel_address: str,
@@ -622,7 +622,7 @@ class Client(ABC):
             check_against_pd=check_against_pd,
         )
 
-    @service()
+    @async_inspector()
     async def get_paramset(
         self,
         address: str,
@@ -648,7 +648,7 @@ class Client(ABC):
                 f"GET_PARAMSET failed with for {address}/{paramset_key}: {reduce_args(args=ex.args)}"
             ) from ex
 
-    @service(measure_performance=True)
+    @async_inspector(measure_performance=True)
     async def put_paramset(
         self,
         channel_address: str,
@@ -803,7 +803,7 @@ class Client(ABC):
             return parameter_data["TYPE"]
         return None
 
-    @service(re_raise=False)
+    @async_inspector(re_raise=False)
     async def fetch_paramset_description(self, channel_address: str, paramset_key: ParamsetKey) -> None:
         """Fetch a specific paramset and add it to the known ones."""
         _LOGGER.debug("FETCH_PARAMSET_DESCRIPTION: %s for %s", paramset_key, channel_address)
@@ -818,7 +818,7 @@ class Client(ABC):
                 paramset_description=paramset_description,
             )
 
-    @service(re_raise=False)
+    @async_inspector(re_raise=False)
     async def fetch_paramset_descriptions(self, device_description: DeviceDescription) -> None:
         """Fetch paramsets for provided device description."""
         data = await self.get_paramset_descriptions(device_description=device_description)
@@ -832,7 +832,7 @@ class Client(ABC):
                     paramset_description=paramset_description,
                 )
 
-    @service(re_raise=False, no_raise_return={})
+    @async_inspector(re_raise=False, no_raise_return={})
     async def get_paramset_descriptions(
         self, device_description: DeviceDescription
     ) -> dict[str, dict[ParamsetKey, dict[str, ParameterData]]]:
@@ -866,7 +866,7 @@ class Client(ABC):
             )
         return None
 
-    @service()
+    @async_inspector()
     async def get_all_paramset_descriptions(
         self, device_descriptions: tuple[DeviceDescription, ...]
     ) -> dict[str, dict[ParamsetKey, dict[str, ParameterData]]]:
@@ -876,12 +876,12 @@ class Client(ABC):
             all_paramsets.update(await self.get_paramset_descriptions(device_description=device_description))
         return all_paramsets
 
-    @service()
+    @async_inspector()
     async def has_program_ids(self, channel_hmid: str) -> bool:
         """Return if a channel has program ids."""
         return False
 
-    @service(re_raise=False, measure_performance=True)
+    @async_inspector(re_raise=False, measure_performance=True)
     async def list_devices(self) -> tuple[DeviceDescription, ...] | None:
         """List devices of homematic backend."""
         try:
@@ -894,12 +894,12 @@ class Client(ABC):
             )
         return None
 
-    @service()
+    @async_inspector()
     async def report_value_usage(self, address: str, value_id: str, ref_counter: int) -> bool:
         """Report value usage."""
         return False
 
-    @service()
+    @async_inspector()
     async def update_device_firmware(self, device_address: str) -> bool:
         """Update the firmware of a homematic device."""
         if device := self.central.get_device(address=device_address):
@@ -924,7 +924,7 @@ class Client(ABC):
             return result
         return False
 
-    @service(re_raise=False)
+    @async_inspector(re_raise=False)
     async def update_paramset_descriptions(self, device_address: str) -> None:
         """Update paramsets descriptions for provided device_address."""
         if not self.central.device_descriptions.get_device_descriptions(interface_id=self.interface_id):
@@ -973,7 +973,7 @@ class ClientCCU(Client):
         """Return the supports_ping_pong info of the backend."""
         return True
 
-    @service(re_raise=False, measure_performance=True)
+    @async_inspector(re_raise=False, measure_performance=True)
     async def fetch_device_details(self) -> None:
         """Get all names via JSON-RPS and store in data.NAMES."""
         if json_result := await self._json_rpc_client.get_device_details():
@@ -993,7 +993,7 @@ class ClientCCU(Client):
         else:
             _LOGGER.debug("FETCH_DEVICE_DETAILS: Unable to fetch device details via JSON-RPC")
 
-    @service(re_raise=False, measure_performance=True)
+    @async_inspector(re_raise=False, measure_performance=True)
     async def fetch_all_device_data(self) -> None:
         """Fetch all device data from CCU."""
         try:
@@ -1017,7 +1017,7 @@ class ClientCCU(Client):
             self.interface,
         )
 
-    @service(re_raise=False, no_raise_return=False)
+    @async_inspector(re_raise=False, no_raise_return=False)
     async def check_connection_availability(self, handle_ping_pong: bool) -> bool:
         """Check if _proxy is still initialized."""
         try:
@@ -1042,22 +1042,22 @@ class ClientCCU(Client):
         self.modified_at = INIT_DATETIME
         return False
 
-    @service()
+    @async_inspector()
     async def execute_program(self, pid: str) -> bool:
         """Execute a program on CCU."""
         return await self._json_rpc_client.execute_program(pid=pid)
 
-    @service()
+    @async_inspector()
     async def set_program_state(self, pid: str, state: bool) -> bool:
         """Set the program state on CCU."""
         return await self._json_rpc_client.set_program_state(pid=pid, state=state)
 
-    @service()
+    @async_inspector()
     async def has_program_ids(self, channel_hmid: str) -> bool:
         """Return if a channel has program ids."""
         return await self._json_rpc_client.has_program_ids(channel_hmid=channel_hmid)
 
-    @service()
+    @async_inspector()
     async def report_value_usage(self, address: str, value_id: str, ref_counter: int) -> bool:
         """Report value usage."""
         try:
@@ -1067,34 +1067,34 @@ class ClientCCU(Client):
                 f"REPORT_VALUE_USAGE failed with: {address}/{value_id}/{ref_counter}: {reduce_args(args=ex.args)}"
             ) from ex
 
-    @service(measure_performance=True)
+    @async_inspector(measure_performance=True)
     async def set_system_variable(self, legacy_name: str, value: Any) -> bool:
         """Set a system variable on CCU / Homegear."""
         return await self._json_rpc_client.set_system_variable(legacy_name=legacy_name, value=value)
 
-    @service()
+    @async_inspector()
     async def delete_system_variable(self, name: str) -> bool:
         """Delete a system variable from CCU / Homegear."""
         return await self._json_rpc_client.delete_system_variable(name=name)
 
-    @service()
+    @async_inspector()
     async def get_system_variable(self, name: str) -> Any:
         """Get single system variable from CCU / Homegear."""
         return await self._json_rpc_client.get_system_variable(name=name)
 
-    @service(re_raise=False)
+    @async_inspector(re_raise=False)
     async def get_all_system_variables(
         self, markers: tuple[DescriptionMarker | str, ...]
     ) -> tuple[SystemVariableData, ...] | None:
         """Get all system variables from CCU."""
         return await self._json_rpc_client.get_all_system_variables(markers=markers)
 
-    @service(re_raise=False)
+    @async_inspector(re_raise=False)
     async def get_all_programs(self, markers: tuple[DescriptionMarker | str, ...]) -> tuple[ProgramData, ...]:
         """Get all programs, if available."""
         return await self._json_rpc_client.get_all_programs(markers=markers)
 
-    @service(re_raise=False, no_raise_return={})
+    @async_inspector(re_raise=False, no_raise_return={})
     async def get_all_rooms(self) -> dict[str, set[str]]:
         """Get all rooms from CCU."""
         rooms: dict[str, set[str]] = {}
@@ -1106,7 +1106,7 @@ class ClientCCU(Client):
                 rooms[address].update(names)
         return rooms
 
-    @service(re_raise=False, no_raise_return={})
+    @async_inspector(re_raise=False, no_raise_return={})
     async def get_all_functions(self) -> dict[str, set[str]]:
         """Get all functions from CCU."""
         functions: dict[str, set[str]] = {}
@@ -1130,7 +1130,7 @@ class ClientJsonCCU(ClientCCU):
         """Init the client."""
         self._system_information = await self._get_system_information()
 
-    @service(re_raise=False, no_raise_return=False)
+    @async_inspector(re_raise=False, no_raise_return=False)
     async def check_connection_availability(self, handle_ping_pong: bool) -> bool:
         """Check if proxy is still initialized."""
         return await self._json_rpc_client.is_present(interface=self.interface)
@@ -1140,7 +1140,7 @@ class ClientJsonCCU(ClientCCU):
         """Return the supports_ping_pong info of the backend."""
         return False
 
-    @service(re_raise=False)
+    @async_inspector(re_raise=False)
     async def get_device_description(self, device_address: str) -> DeviceDescription | None:
         """Get device descriptions from CCU / Homegear."""
         try:
@@ -1152,7 +1152,7 @@ class ClientJsonCCU(ClientCCU):
             _LOGGER.warning("GET_DEVICE_DESCRIPTIONS failed: %s [%s]", ex.name, reduce_args(args=ex.args))
         return None
 
-    @service()
+    @async_inspector()
     async def get_paramset(
         self,
         address: str,
@@ -1183,7 +1183,7 @@ class ClientJsonCCU(ClientCCU):
                 f"GET_PARAMSET failed with for {address}/{paramset_key}: {reduce_args(args=ex.args)}"
             ) from ex
 
-    @service(log_level=logging.NOTSET)
+    @async_inspector(log_level=logging.NOTSET)
     async def get_value(
         self,
         channel_address: str,
@@ -1221,7 +1221,7 @@ class ClientJsonCCU(ClientCCU):
                 f"GET_VALUE failed with for: {channel_address}/{parameter}/{paramset_key}: {reduce_args(args=ex.args)}"
             ) from ex
 
-    @service(re_raise=False, measure_performance=True)
+    @async_inspector(re_raise=False, measure_performance=True)
     async def list_devices(self) -> tuple[DeviceDescription, ...] | None:
         """List devices of homematic backend."""
         try:
@@ -1342,12 +1342,12 @@ class ClientHomegear(Client):
         """Return the supports_ping_pong info of the backend."""
         return False
 
-    @service(re_raise=False)
+    @async_inspector(re_raise=False)
     async def fetch_all_device_data(self) -> None:
         """Fetch all device data from CCU."""
         return
 
-    @service(re_raise=False, measure_performance=True)
+    @async_inspector(re_raise=False, measure_performance=True)
     async def fetch_device_details(self) -> None:
         """Get all names from metadata (Homegear)."""
         _LOGGER.debug("FETCH_DEVICE_DETAILS: Fetching names via Metadata")
@@ -1365,7 +1365,7 @@ class ClientHomegear(Client):
                     address,
                 )
 
-    @service(re_raise=False, no_raise_return=False)
+    @async_inspector(re_raise=False, no_raise_return=False)
     async def check_connection_availability(self, handle_ping_pong: bool) -> bool:
         """Check if proxy is still initialized."""
         try:
@@ -1390,24 +1390,24 @@ class ClientHomegear(Client):
         """Set the program state on Homegear."""
         return True
 
-    @service(measure_performance=True)
+    @async_inspector(measure_performance=True)
     async def set_system_variable(self, legacy_name: str, value: Any) -> bool:
         """Set a system variable on CCU / Homegear."""
         await self._proxy.setSystemVariable(legacy_name, value)
         return True
 
-    @service()
+    @async_inspector()
     async def delete_system_variable(self, name: str) -> bool:
         """Delete a system variable from CCU / Homegear."""
         await self._proxy.deleteSystemVariable(name)
         return True
 
-    @service()
+    @async_inspector()
     async def get_system_variable(self, name: str) -> Any:
         """Get single system variable from CCU / Homegear."""
         return await self._proxy.getSystemVariable(name)
 
-    @service(re_raise=False)
+    @async_inspector(re_raise=False)
     async def get_all_system_variables(
         self, markers: tuple[DescriptionMarker | str, ...]
     ) -> tuple[SystemVariableData, ...] | None:
@@ -1418,17 +1418,17 @@ class ClientHomegear(Client):
                 variables.append(SystemVariableData(vid=name, legacy_name=name, value=value))
         return tuple(variables)
 
-    @service(re_raise=False)
+    @async_inspector(re_raise=False)
     async def get_all_programs(self, markers: tuple[DescriptionMarker | str, ...]) -> tuple[ProgramData, ...] | None:
         """Get all programs, if available."""
         return ()
 
-    @service(re_raise=False, no_raise_return={})
+    @async_inspector(re_raise=False, no_raise_return={})
     async def get_all_rooms(self) -> dict[str, set[str]]:
         """Get all rooms from Homegear."""
         return {}
 
-    @service(re_raise=False, no_raise_return={})
+    @async_inspector(re_raise=False, no_raise_return={})
     async def get_all_functions(self) -> dict[str, set[str]]:
         """Get all functions from Homegear."""
         return {}