+++ hahomematic/model/generic/button.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from hahomematic.const import DataPointCategory
-from hahomematic.decorators import service
+from hahomematic.decorators import async_inspector
 from hahomematic.model.generic.data_point import GenericDataPoint
 
 
@@ -17,7 +17,7 @@ class DpButton(GenericDataPoint[None, bool]):
     _category = DataPointCategory.BUTTON
     _validate_state_change = False
 
-    @service()
+    @async_inspector()
     async def press(self) -> None:
         """Handle the button press."""
         await self.send_value(value=True)