+++ hahomematic/model/hub/switch.py
@@ -3,7 +3,7 @@
 from __future__ import annotations
 
 from hahomematic.const import DataPointCategory
-from hahomematic.decorators import service
+from hahomematic.decorators import async_inspector
 from hahomematic.model.decorators import state_property
 from hahomematic.model.hub.data_point import GenericProgramDataPoint, GenericSysvarDataPoint
 
@@ -25,13 +25,13 @@ class ProgramDpSwitch(GenericProgramDataPoint):
         """Get the value of the data_point."""
         return self._is_active
 
-    @service()
+    @async_inspector()
     async def turn_on(self) -> None:
         """Turn the program on."""
         await self.central.set_program_state(pid=self._pid, state=True)
         await self._central.fetch_program_data(scheduled=False)
 
-    @service()
+    @async_inspector()
     async def turn_off(self) -> None:
         """Turn the program off."""
         await self.central.set_program_state(pid=self._pid, state=False)