+++ changelog.md
@@ -1,6 +1,7 @@
 # Version 2025.1.4 (2025-01-09)
 
 - Cleanup: Use defaultdict, improve naming
+- Rename decorator @service to @async\_/@sync_inspector
 
 # Version 2025.1.3 (2025-01-08)